requirements.txt
# Core Dependencies
python-dotenv==1.0.0
cloudflare==2.11.1
pyyaml==6.0.1
typing-extensions==4.9.0

# Deployment and Validation
wrangler==3.0.1

# Development and Testing
pytest==7.4.3
mypy==1.7.1
flake8==6.1.0
bandit==1.7.5

# Logging and Reporting
structlog==23.1.0

# Template Generator Dependencies
jsonschema==4.17.3
pyyaml==6.0.1
typing-extensions==4.9.0
jinja2==3.1.2
click==8.1.3  # For CLI support
MarkupSafe==2.1.1

# Security and Validation
jsonschema==4.17.3

# Optional: Additional utilities
click==8.1.7
rich==10.16.2
